+++ src/main/java/app/AppBuilder.java
@@ -7,13 +7,21 @@ import java.time.Instant;
 import constants.Constants;
 import dataaccessinterface.BookmarkedLocationStorage;
 import dataaccessobjects.InDiskBookmarkStorage;
-import dataaccessobjects.OkHttpsPointWeatherGatewayXml;
 
 import entity.ProgramTime;
 import entity.Viewport;
 import entity.Location;
 import interfaceadapter.bookmark.visitbookmark.VisitBookmarkController;
 import interfaceadapter.bookmark.visitbookmark.VisitBookmarkPresenter;
+import interfaceadapter.weatherlayers.legend.LegendPresenter;
+import interfaceadapter.weatherlayers.legend.LegendViewModel;
+import interfaceadapter.weatherlayers.updateoverlay.UpdateOverlayController;
+import interfaceadapter.weatherlayers.updateoverlay.UpdateOverlayPresenter;
+import interfaceadapter.weatherlayers.updateoverlay.UpdateOverlaySizeController;
+import interfaceadapter.weatherlayers.updateoverlay.UpdateOverlayViewModel;
+import interfaceadapter.weatherlayers.layers.WeatherLayersController;
+import interfaceadapter.weatherlayers.layers.WeatherLayersPresenter;
+import interfaceadapter.weatherlayers.layers.WeatherLayersViewModel;
 import usecase.bookmark.visitbookmark.VisitBookmarkInputBoundary;
 import usecase.bookmark.visitbookmark.VisitBookmarkOutputBoundary;
 import usecase.bookmark.visitbookmark.VisitBookmarkUseCase;
@@ -24,13 +32,9 @@ import interfaceadapter.bookmark.listbookmark.ListBookmarksController;
 import interfaceadapter.bookmark.listbookmark.ListBookmarksPresenter;
 import interfaceadapter.bookmark.removebookmark.RemoveBookmarkController;
 import interfaceadapter.bookmark.removebookmark.RemoveBookmarkPresenter;
-import interfaceadapter.infopanel.InfoPanelController;
-import interfaceadapter.infopanel.InfoPanelPresenter;
-import interfaceadapter.infopanel.InfoPanelViewModel;
 import interfaceadapter.maptime.programtime.ProgramTimeController;
 import interfaceadapter.maptime.programtime.ProgramTimePresenter;
 import interfaceadapter.maptime.timeanimation.TimeAnimationController;
-import interfaceadapter.weatherlayers.*;
 import org.openstreetmap.gui.jmapviewer.JMapViewer;
 import usecase.bookmark.addbookmark.AddBookmarkInputBoundary;
 import usecase.bookmark.addbookmark.AddBookmarkOutputBoundary;
@@ -40,24 +44,22 @@ import usecase.bookmark.listbookmark.ListBookmarksUseCase;
 import usecase.bookmark.removebookmark.RemoveBookmarkInputBoundary;
 import usecase.bookmark.removebookmark.RemoveBookmarkOutputBoundary;
 import usecase.bookmark.removebookmark.RemoveBookmarkUseCase;
-import usecase.infopanel.InfoPanelInteractor;
-import usecase.infopanel.PointWeatherFetcher;
 import usecase.maptime.UpdateMapTimeInputBoundary;
 import usecase.weatherlayers.layers.*;
-import usecase.weatherlayers.update.UpdateOverlayOutputBoundary;
-import usecase.weatherlayers.update.UpdateOverlaySizeUseCase;
-import usecase.weatherlayers.update.UpdateOverlayUseCase;
+import usecase.weatherlayers.updateoverlay.UpdateOverlayOutputBoundary;
+import usecase.weatherlayers.updateoverlay.UpdateOverlaySizeUseCase;
+import usecase.weatherlayers.updateoverlay.UpdateOverlayUseCase;
 import usecase.maptime.UpdateMapTimeOutputBoundary;
 import usecase.maptime.UpdateMapTimeUseCase;
 import view.*;
 import interfaceadapter.maptime.programtime.ProgramTimeViewModel;
 import dataaccessobjects.CachedTileRepository;
 import entity.OverlayManager;
-import interfaceadapter.mapinteraction.MapViewModel;
-import interfaceadapter.mapinteraction.PanAndZoomController;
-import interfaceadapter.mapinteraction.PanAndZoomPresenter;
-import usecase.mapinteraction.PanAndZoomUseCase;
-import usecase.mapinteraction.PanAndZoomInputBoundary;
+import interfaceadapter.mapnavigation.MapViewModel;
+import interfaceadapter.mapnavigation.PanAndZoomController;
+import interfaceadapter.mapnavigation.PanAndZoomPresenter;
+import usecase.mapnavigation.PanAndZoomUseCase;
+import usecase.mapnavigation.PanAndZoomInputBoundary;
 import dataaccessinterface.SavedMapOverlaySettings;
 import dataaccessobjects.InDiskMapOverlaySettingsStorage;
 import interfaceadapter.mapsettings.loadmapsettings.AutoLoadMapSettingsPresenter;
