+++ src/main/java/interfaceadapter/mapnavigation/MapViewModel.java
@@ -1,4 +1,4 @@
-package interfaceadapter.mapinteraction;
+package interfaceadapter.mapnavigation;
 
 import entity.Location;
 import java.beans.PropertyChangeSupport;
