+++ src/main/java/interfaceadapter/weatherlayers/legend/LegendViewModel.java
@@ -1,4 +1,4 @@
-package interfaceadapter.weatherlayers;
+package interfaceadapter.weatherlayers.legend;
 
 import interfaceadapter.ViewModel;
