+++ src/main/java/interfaceadapter/weatherlayers/updateoverlay/UpdateOverlayViewModel.java
@@ -1,7 +1,7 @@
-package interfaceadapter.weatherlayers;
+package interfaceadapter.weatherlayers.updateoverlay;
 
 import interfaceadapter.ViewModel;
-
+import interfaceadapter.weatherlayers.layers.OverlayState;
 
 
 public class UpdateOverlayViewModel extends ViewModel<OverlayState> {
