+++ src/main/java/usecase/mapnavigation/PanAndZoomInputBoundary.java
@@ -1,4 +1,4 @@
-package usecase.mapinteraction;
+package usecase.mapnavigation;
 
 public interface PanAndZoomInputBoundary {
     void updateViewport(PanAndZoomInputData input);
