+++ src/main/java/usecase/mapnavigation/PanAndZoomOutputBoundary.java
@@ -1,4 +1,4 @@
-package usecase.mapinteraction;
+package usecase.mapnavigation;
 
 public interface PanAndZoomOutputBoundary {
     void present(PanAndZoomOutputData outputData);
