+++ src/main/java/usecase/mapnavigation/ZoomOutOfBoundsException.java
@@ -1,4 +1,4 @@
-package usecase.mapinteraction;
+package usecase.mapnavigation;
 
 public class ZoomOutOfBoundsException extends Exception {
     public ZoomOutOfBoundsException(String message) {
