+++ src/main/java/usecase/weatherlayers/updateoverlay/UpdateOverlayOutputBoundary.java
@@ -1,4 +1,4 @@
-package usecase.weatherlayers.update;
+package usecase.weatherlayers.updateoverlay;
 
 public interface UpdateOverlayOutputBoundary {
     void updateImage(UpdateOverlayOutputData data);
