+++ src/main/java/usecase/weatherlayers/updateoverlay/UpdateOverlaySizeInputBoundary.java
@@ -1,4 +1,4 @@
-package usecase.weatherlayers.update;
+package usecase.weatherlayers.updateoverlay;
 
 public interface UpdateOverlaySizeInputBoundary {
     public void changeSize(UpdateOverlaySizeInputData data);
