+++ src/main/java/usecase/weatherlayers/updateoverlay/UpdateOverlaySizeInputData.java
@@ -1,4 +1,4 @@
-package usecase.weatherlayers.update;
+package usecase.weatherlayers.updateoverlay;
 
 import java.awt.*;
